progress

+++ SocialNetworkAnalysis-NoSQL/src/main/java/org/example/socialnetworkanalysisnosql/services/ProductService.java
@@ -17,7 +17,7 @@ public class ProductService {
     public ProductService() {
     }
 
-    public Product createProduct(String s, String description, long price) {
+    public Product createProduct(String s, String description, double price) {
         Product product = new Product(s, description, price);
         repo.save(product);
         return product;
@@ -39,4 +39,5 @@ public class ProductService {
     public List<Product> findAll() {
         return this.repo.findAll();
     }
+
 }
